requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=2.0.0
pydantic>=1.8.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
python-multipart>=0.0.5,<0.0.7
email-validator>=1.1.3,<2.0.0
